successfully refactored
refactored some code: moved the children of Page to the variable manager. any more children of Page will have to be defined in the variable manager. got a button to finally switch to another page, though there is still a bug where it takes about two seconds to do so (probably the SoundManager initializing). will be fixed in the next commit.